For reusing paths I created additional file with import paths from ./paths
