+++ src/Components/OurWork/OurWork.tsx
@@ -1,0 +1,7 @@
+import './style.scss';
+
+const OurWork = () => {
+  return <div></div>;
+};
+
+export default OurWork;
