+++ src/Components/OurWork/index.ts
@@ -1,0 +1,1 @@
+export { default as OurWork } from './OurWork';
